airbloc: Refactor shared/blockchain (2) * Rename pacakge import path
